# backend_fastapi/openapi_schema.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AI Interviewer API",
    "description": "Простой API для проведения технических интервью по JavaScript",
    "version": "1.0.0"
  },
  "paths": {
    "/": {
      "get": {
        "summary": "Root",
        "description": "Корневой endpoint",
        "operationId": "root__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/api/interview/start": {
      "post": {
        "summary": "Start Interview",
        "description": "Начать новое интервью\n\n- **topic**: Тема интервью (например, \"javascript-basics\")\n- **difficulty**: Уровень сложности (\"junior\", \"middle\", \"senior\")\n- **question_count**: Количество вопросов",
        "operationId": "start_interview_api_interview_start_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InterviewStartRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InterviewSession"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/interview/question": {
      "get": {
        "summary": "Get Question",
        "description": "Получить текущий вопрос для интервью\n\n- **interview_id**: ID сессии интервью",
        "operationId": "get_question_api_interview_question_get",
        "parameters": [
          {
            "name": "interview_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Interview Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/interview/answer": {
      "post": {
        "summary": "Submit Answer",
        "description": "Отправить ответ на вопрос и получить фидбэк\n\n- **interview_id**: ID сессии интервью\n- **question_id**: ID вопроса\n- **answer**: Ответ пользователя\n- **time_spent**: Время, потраченное на ответ (в секундах)",
        "operationId": "submit_answer_api_interview_answer_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnswerRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Feedback"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/interview/end": {
      "post": {
        "summary": "End Interview",
        "description": "Завершить интервью\n\n- **interview_id**: ID сессии интервью",
        "operationId": "end_interview_api_interview_end_post",
        "parameters": [
          {
            "name": "interview_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Interview Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AnswerRequest": {
        "properties": {
          "interview_id": {
            "type": "string",
            "title": "Interview Id"
          },
          "question_id": {
            "type": "string",
            "title": "Question Id"
          },
          "answer": {
            "type": "string",
            "title": "Answer"
          },
          "time_spent": {
            "type": "integer",
            "title": "Time Spent"
          }
        },
        "type": "object",
        "required": [
          "interview_id",
          "question_id",
          "answer",
          "time_spent"
        ],
        "title": "AnswerRequest"
      },
      "Feedback": {
        "properties": {
          "score": {
            "type": "integer",
            "title": "Score"
          },
          "comment": {
            "type": "string",
            "title": "Comment"
          },
          "suggestions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Suggestions"
          },
          "correct_answer": {
            "type": "string",
            "title": "Correct Answer"
          }
        },
        "type": "object",
        "required": [
          "score",
          "comment",
          "suggestions",
          "correct_answer"
        ],
        "title": "Feedback"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "InterviewSession": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "topic": {
            "type": "string",
            "title": "Topic"
          },
          "difficulty": {
            "type": "string",
            "title": "Difficulty"
          },
          "current_question": {
            "type": "integer",
            "title": "Current Question"
          },
          "total_questions": {
            "type": "integer",
            "title": "Total Questions"
          },
          "score": {
            "type": "integer",
            "title": "Score",
            "default": 0
          },
          "start_time": {
            "type": "string",
            "format": "date-time",
            "title": "Start Time"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "id",
          "topic",
          "difficulty",
          "current_question",
          "total_questions",
          "start_time"
        ],
        "title": "InterviewSession"
      },
      "InterviewStartRequest": {
        "properties": {
          "topic": {
            "type": "string",
            "title": "Topic"
          },
          "difficulty": {
            "type": "string",
            "title": "Difficulty",
            "default": "middle"
          },
          "question_count": {
            "type": "integer",
            "title": "Question Count",
            "default": 10
          }
        },
        "type": "object",
        "required": [
          "topic"
        ],
        "title": "InterviewStartRequest"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}